Enhance Android logger functionality with noexcept specifier
- Updated function signatures in `android.h` to include `noexcept` for improved safety and performance.
- Refactored operator overloads and logging implementations to ensure exception safety during logging operations.
- Improved code clarity and maintainability by standardizing function signatures across the Android logging interface.

+++ include/fast_io_hosted/filesystem/dos_at.h
@@ -201,7 +201,7 @@ inline void dos_unlinkat_impl(int dirfd, char const *pathname)
 
 inline constexpr ::std::time_t unix_timestamp_to_time_t(unix_timestamp stmp) noexcept
 {
-	return static_cast<int>(static_cast<::std::time_t>(stmp.seconds));
+	return static_cast<::std::time_t>(stmp.seconds);
 }
 
 inline
